Refactored the sidebar to position ThemeSwitcher and hide sidebar as fixed. Context Menu related components were moved to Menu items
