+++ web-client/src/components/layout/menu/ContextMenu/ContextMenu.component.tsx
@@ -1,4 +1,4 @@
-import { ContextMenuItem } from "@/components/menus"
+import { ContextMenuItem } from "@/components/layout/menu"
 import { useEffect, useRef, useState, type KeyboardEvent } from "react"
 import { Button } from "@/components/forms"
 import { assertIsNode } from "@/utils/helpers/types.helpers"
