+++ web-client/src/components/layout/menu/index.ts
@@ -2,3 +2,5 @@ export { Sidebar } from "./Sidebar/Sidebar.component"
 export { MenuLink } from "./MenuLink/MenuLink.component"
 export { SidebarHeader } from "./SidebarHeader/SidebarHeader.component"
 export { ThemeSwitcher } from "./ThemeSwitcher/ThemeSwitcher.component"
+export { ContextMenu } from "./ContextMenu/ContextMenu.component"
+export { ContextMenuItem } from "./ContextMenuItem/ContextMenuItem.component"
