--- web-client/src/components/menus/index.ts
@@ -1,2 +1,0 @@
-export { ContextMenu } from "@/components/menus/ContextMenu/ContextMenu.component"
-export { ContextMenuItem } from "@/components/menus/ContextMenuItem/ContextMenuItem.component"
